--- a/plonky2/src/hash/arch/x86_64/mod.rs
+++ b/plonky2/src/hash/arch/x86_64/mod.rs
@@ -1,6 +1,6 @@
 // // Requires:
 // // - AVX2
 #[cfg(target_feature = "avx2")]
-pub(crate) mod poseidon_goldilocks_avx2;
-#[cfg(target_feature = "avx2")]
 pub(crate) mod goldilocks_avx2;
+#[cfg(target_feature = "avx2")]
+pub(crate) mod poseidon_goldilocks_avx2;