… plonky2/src/hash/arch/x86_64/poseidon_goldilocks_avx2.rs
… plonky2/src/hash/arch/x86_64/poseidon_goldilocks_avx2_bmi2.rs